refactor(cli): Rename scripts→cli, integrate defopt
- New `trackers/cli/` package: defopt-based entry point, `--tracker-params key=val` replaces per-param `--tracker.<param>` flags, typed function signatures on all four subcommands
- `trackers/scripts/` converted to thin backward-compat re-export shims
- Tests moved from `tests/scripts/` to `tests/cli/`; 75 tests ported/rewritten for defopt API; tests/scripts stubs removed
- Add `defopt>=7.0.0` to dependencies (pulls docutils + sphinxcontrib-napoleon + pockets)
- Console-scripts entry updated to `trackers.cli.__main__:main`

---
