+++ pyproject.toml
@@ -41,6 +41,7 @@ dependencies = [
     "opencv-python>=4.8.0",
     "rich>=13.0.0",
     "requests>=2.28.0",
+    "defopt>=7.0.0",
 ]
 
 [project.optional-dependencies]
@@ -48,7 +49,7 @@ detection = ["inference-models>=0.19.0"]
 tune = ["optuna>=3.0.0"]
 
 [project.scripts]
-trackers = "trackers.scripts.__main__:main"
+trackers = "trackers.cli.__main__:main"
 
 [dependency-groups]
 dev = [
