+++ tests/cli/__init__.py
@@ -1,9 +1,5 @@
-#!/usr/bin/env python
 # ------------------------------------------------------------------------
 # Trackers
 # Copyright (c) 2026 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-
-# Backward-compat shim — use trackers.cli.eval instead.
-from trackers.cli.eval import eval_cmd  # noqa: F401
